+++ src/typedefs.rs
@@ -6,3 +6,24 @@ pub enum CalibrationMode {
     /// Pulses are then applied once per minute.
     Slow = 0x0
 }
+
+pub enum PowerManagement {
+    /// Battery switch-over function is enabled in standard mode; battery low detection function is
+    /// enabled.
+    SwitchOverStandardOnLowDetectionOn = 0x0,
+    /// Battery switch-over function is enabled in direct switching mode; battery low detection
+    /// function is enabled.
+    SwitchOverDirectOnLowDetectionOn = 0x1,
+    /// Battery switch-over function is disabled - only one power supply (VDD); battery low
+    /// detection function is enabled.
+    SwitchOverOffLowDetectionOn = 0x2,
+    /// Battery switch-over function is enabled in standard mode; battery low detection function is
+    /// disabled.
+    SwitchOverStandardOnLowDetectionOff = 0x4,
+    /// Battery switch-over function is enabled in direct switching mode; battery low detection
+    /// function is disabled.
+    SwitchOverDirectOnLowDetectionOff = 0x5,
+    /// Battery switch-over function is disabled - only one power supply (VDD); battery low
+    /// detection function is disabled.
+    SwitchOverOffLowDetectionOff = 0x7
+}